웹소켓, 30m 이내의 퀘스트 반환 기능 반환 message타입 미해결

+++ src/main/java/com/hybrid/gungduk/dao/AcheiveDao.java
@@ -24,7 +24,7 @@ public class AcheiveDao {
 		return sqlSession.selectOne("acheive.countYes", map);
 	}
 	
-	public AcheiveDto putData(double GBacheive, double CGacheive, double DSacheive, double CDacheive){
+	public AcheiveDto putData(int GBacheive, int CGacheive, int DSacheive, int CDacheive){
 		AcheiveDto acheiveDto = new AcheiveDto();
 		
 		acheiveDto.setGyeongbok(GBacheive);
